requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
datasets>=2.14.0
pillow>=9.0.0
accelerate>=0.20.0
wandb>=0.15.0
fastapi>=0.100.0
uvicorn>=0.23.0
numpy>=1.24.0
tqdm>=4.65.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
ipywidgets>=8.0.0
requests>=2.31.0
packaging>=23.0
tokenizers>=0.14.0
safetensors>=0.3.0
huggingface_hub>=0.16.0
peft>=0.5.0
bitsandbytes>=0.41.0